requirements.txt
fastapi
azure-functions
scikit-surprise
pandas
uvicorn
pydantic
httpx
jinja2
python-multipart
###########
fastapi
uvicorn
gunicorn
pandas
numpy
scikit-learn
scikit-surprise
jinja2
python-multipart